css,js 파일 모듈화 해서 중복함수,중복데이터,중복css 분리함
