+++ login.html
@@ -30,7 +30,7 @@
     <!-- End Google Tag Manager (noscript) -->
    
   <main class="login-layer">
-    <section class="login-section1">
+    <section class=".login-section1-layer">
         <a  href="/index.html"
             target="_self">
             <img 
@@ -40,11 +40,11 @@
             />   
         </a>
     </section>
-    <section class="login-section2">
+    <section class=".login-section2-layer">
       
       <form class="form">
         
-        <div class="login-form-control">
+        <div class="form-control">
           <label for="usermail">
             이메일
           </label>
@@ -55,7 +55,7 @@
           <small>Error message</small>
         </div>
 
-        <div class="login-form-control">
+        <div class="form-control">
           <label for="password">
             비밀번호
           </label>
@@ -120,8 +120,8 @@
       </a>
       </div>
     </section>
-  </main>
-  <!-- <script src="../javascript/login.js"></script> -->
-  <script type="text/javascript" src="/javascript/login.js"></script> 
+  </main>  
+  <!-- <script type="text/javascript" src="/javascript/login.js"></script>  -->
+  <script type="module" src="/module-js/login.mjs"></script> 
 </body>
 </html>
