+++ styles/signup.css
@@ -20,17 +20,6 @@
   /* gap:16px; */    
 }
 
-#signup-panda-logo {
-  width: 396px;  
-  /* width: 61.87%;  */
-  height:132px;
-}
-
-#signup-panda-logo:hover {
-  /* background-color: #ffffff;
-  color: #000000; */
-  cursor:pointer;
-}
 
 .signup-section2-layer{  
   /* background-color: #33f10d;  */
@@ -38,136 +27,6 @@
   /* height: 548px; */
 }
 
-.form {
-  /* background-color:burlywood; */  
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap:24px; 
-  
-  /* justify-content:space-around; */
-  /* align-items:center; */
-  margin: 16px auto;
-}
-
-.signup-form-control {
-  /* background-color: blueviolet; */
-  width:100%;
-  display: flex;
-  flex-direction: column;
-  gap:12px; 
-}
-
-.signup-form-control label{    
-  color: #1F2937;  
-  font-size: 1.125rem;  
-  line-height: 1.2rem;  
-  font-weight: 700;
-  text-align:start;
-
-  /* width:100%; */
-  /* height: 42px; */
-  /* gap:10px; */
-  /* font-size: 18px;
-  line-height: 21.48px; */
-  /* line-height: 1.343rem; */  
-  /* display: block; */  
-  /* background-color: yellow; */
-}
-
-.signup-wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  /* background-color: red; */
-}
-
-.password-toggle-button {
-  position: absolute;
-  right: 24px;
-  /* background-color: green; */
-}
-
-.signup-form-control input {  
-  width: 100%;  
-  border-radius:12px;
-  padding:16px 24px 16px 24px;
-  background-color: #F3F4F6;
-  color: #9CA3AF;
-  font-weight: 400;
-  font-size: 1rem;
-  line-height: 1.2rem;
-  text-align:start;
-  border-width:0;  
-
-  /* height: 56px; */
-  /* gap:10px; */
-  /* display: block; */
-  /* font-size: 16px;
-  line-height: 26px; */
-  /* line-height: 1.3rem;*/
-}
-
-/* .signup-input:placeholder-shown + .signup-button{
-    background-color: #9CA3AF;
-  }   */
-
-.signup-form-control input:focus {
-  border-width: 1px;
-  border-color: #3692FF;
-  outline: none;  
-}
-
-.signup-form-control.success input {
-  border-color: var(--success-color);
-}
-
-.signup-form-control.error input {
-  border-color: var(--error-color);
-}
-
-.signup-form-control small {
-  color: var(--error-color);
-  position: relative;
-  bottom: 8px;
-  left: 0;
-  visibility: hidden;
-  /* visibility: visible; */
-}
-
-.signup-form-control.error small {
-  visibility: visible;
-}
-
-
-.signup-button {  
-  width: 100%;  
-  border-radius:40px;
-  padding:16px 124px 16px 124px;
-  background-color: #9CA3AF;
-  outline: none;    
-  color: #F3F4F6;
-  font-size: 1.25rem;
-  /* 버튼이 섹션을 넘어가서 1.2rem으로 줄임 */  
-  line-height: 1.2rem;
-  text-align:center;
-  border-width:0;
-  cursor:pointer;
-
-  /* height: 56px; */
-  /* gap:10px;   */  
-  /* background-color: #3692FF; */
-  /* display: block; */
-  /* font-size: 20px;
-  line-height: 24px; */
-  /* font-weight: 600; */
-}
-
-.signup-button:hover { 
-  background-color: #3692FF;
-  /* background-color: #1967d6; 더 진한 파랑색 */  
-  outline: none;
-}
 
 
 .signup-section3-layer{
